+++ src/distr/weighted/mod.rs
@@ -28,6 +28,9 @@ pub trait Weight: Clone {
 
     /// Checked addition
     ///
+    /// Note that for floating-point formats with a representation of infinity,
+    /// overflow-to-infinity is not considered an error.
+    ///
     /// -   `Result::Ok`: On success, `v` is added to `self`
     /// -   `Result::Err`: Returns an error when `Self` cannot represent the
     ///     result of `self + v` (i.e. overflow). The value of `self` should be